public project generator

+++ core/src/demo/xmake.sh
@@ -60,6 +60,8 @@ target "demo"
 xmake_after_install() {
     local target=${1}
     local installdir=${2}
-    print "\t@if test -f ${installdir}/bin/xmake.exe; then rm ${installdir}/bin/xmake.exe; fi" >> "${xmake_sh_makefile}"
-    print "\t@cp ${buildir}/xmake.exe ${installdir}/share/xmake" >> "${xmake_sh_makefile}"
+    if test_eq "${project_generator}" "gmake"; then
+        print "\t@if test -f ${installdir}/bin/xmake.exe; then rm ${installdir}/bin/xmake.exe; fi" >> "${xmake_sh_makefile}"
+        print "\t@cp ${buildir}/xmake.exe ${installdir}/share/xmake" >> "${xmake_sh_makefile}"
+    fi
 }
